#1486 Add option to show articles by category section
