+++ next/src/components/sections/ArticlesListSection/ArticlesAllSection.tsx
@@ -1,0 +1,114 @@
+import { Typography } from '@bratislava/component-library'
+import { keepPreviousData, useQuery } from '@tanstack/react-query'
+import React from 'react'
+
+import BlogPostCard from '@/src/components/cards/BlogPostCard'
+import BlogPostsFilter from '@/src/components/common/BlogPostsFilter/BlogPostsFilter'
+import Pagination from '@/src/components/common/Pagination/Pagination'
+import { ArticlesSectionFragment } from '@/src/services/graphql'
+import { client } from '@/src/services/graphql/gql'
+import {
+  articlesDefaultFilters,
+  articlesFetcher,
+  getArticlesQueryKey,
+} from '@/src/services/meili/fetchers/articlesFetcher'
+import { getCategoryColorLocalStyle } from '@/src/utils/colors'
+import { getNumericLocalDate } from '@/src/utils/formatDate'
+import { generateImageSizes } from '@/src/utils/generateImageSizes'
+import { useLocale } from '@/src/utils/useLocale'
+import { useRoutePreservedState } from '@/src/utils/useRoutePreservedState'
+import { useTranslation } from '@/src/utils/useTranslation'
+
+const imageSizes = generateImageSizes({ default: '100vw', md: '50vw', lg: '33vw' })
+
+type Props = {
+  section: ArticlesSectionFragment
+}
+
+const ArticlesAllSection = ({ section }: Props) => {
+  const { t } = useTranslation()
+  const locale = useLocale()
+
+  const { title, text } = section
+
+  const [filters, setFilters] = useRoutePreservedState({
+    ...articlesDefaultFilters,
+  })
+
+  const { data } = useQuery({
+    queryKey: getArticlesQueryKey(filters, locale),
+    queryFn: () => articlesFetcher(filters, locale),
+    placeholderData: keepPreviousData,
+  })
+
+  const { data: pageCategoriesData } = useQuery({
+    queryKey: ['PageCategories', locale],
+    queryFn: () => client.PageCategories({ locale }),
+    staleTime: Infinity,
+  })
+
+  const { data: tagsData } = useQuery({
+    queryKey: ['Tags', locale],
+    queryFn: () => client.Tags({ locale }),
+    staleTime: Infinity,
+  })
+
+  const handlePageChange = (page: number) => {
+    setFilters({ ...filters, page })
+  }
+
+  const handleTagsChange = (tags: string[]) => {
+    setFilters({ ...filters, tagIds: tags })
+  }
+
+  return (
+    <div className="flex flex-col gap-8">
+      <BlogPostsFilter
+        pageCategories={pageCategoriesData?.pageCategories?.data ?? []}
+        blogPostsTags={tagsData?.tags?.data ?? []}
+        onTagChange={handleTagsChange}
+      />
+      {title || text ? (
+        <div className="flex flex-col gap-2">
+          {title && <Typography type="h2">{title}</Typography>}
+          {text && <Typography type="p">{text}</Typography>}
+        </div>
+      ) : null}
+      <div className="grid grid-cols-1 gap-8 sm:grid-cols-2 lg:grid-cols-3">
+        {data?.hits.map((card) => {
+          if (!card.attributes) return null
+
+          // TODO refactor sections that use BlogPostCard - it needs too much duplicate code while passing props
+          const { title: blogPostTitle, slug, coverMedia, addedAt, perex, tag } = card.attributes
+          const tagColor = tag?.data?.attributes?.pageCategory?.data?.attributes?.color
+          const tagTitle = tag?.data?.attributes?.title
+
+          return (
+            <BlogPostCard
+              key={slug}
+              style={getCategoryColorLocalStyle({ color: tagColor })}
+              date={getNumericLocalDate(addedAt)}
+              tag={tagTitle ?? undefined}
+              title={blogPostTitle ?? ''}
+              text={perex ?? undefined}
+              linkProps={{ children: t('readMore'), href: `/spravy/${slug}` }}
+              imgSrc={coverMedia?.data?.attributes?.url}
+              imgSizes={imageSizes}
+            />
+          )
+        })}
+      </div>
+
+      {data?.estimatedTotalHits ? (
+        <Pagination
+          key={filters.search}
+          totalCount={Math.ceil(data.estimatedTotalHits / filters.pageSize)}
+          currentPage={filters.page}
+          onPageChange={handlePageChange}
+        />
+      ) : null}
+    </div>
+  )
+}
+
+export default ArticlesAllSection
