+++ strapi/src/components/sections/articles.json
@@ -11,6 +11,15 @@
     },
     "text": {
       "type": "text"
+    },
+    "category": {
+      "type": "relation",
+      "relation": "oneToOne",
+      "target": "api::page-category.page-category"
+    },
+    "showAll": {
+      "type": "boolean",
+      "default": false
     }
   }
 }
